refactor: split packages/agent/daemon/runtime/standard_acp_adapter_test.go
Flow-Cycle: 3ae1fccd-e7e1-482f-90c1-0c0998dbaab7
