+++ packages/agent/daemon/runtime/standard_acp_interactive_test.go
@@ -1,0 +1,499 @@
+package agentruntime
+
+import (
+	"context"
+	"errors"
+	"strings"
+	"sync"
+	"testing"
+	"time"
+
+	activityshared "github.com/tutti-os/tutti/packages/agent/daemon/activity/events"
+)
+
+func TestStandardACPAdapterSessionStateExposesPendingAskUserPrompt(t *testing.T) {
+	t.Parallel()
+
+	transport := newStandardACPTransport("Hermes Agent", "hermes-session-interactive-1")
+	transport.conn.promptKind = "ask-user"
+	adapter := newHermesExtensionTestAdapter(transport)
+	session := standardTestSession(hermesExtensionTestProvider)
+	if _, err := adapter.Start(context.Background(), session); err != nil {
+		t.Fatalf("Start: %v", err)
+	}
+	session.ProviderSessionID = "hermes-session-interactive-1"
+
+	var mu sync.Mutex
+	var emittedActivity []activityshared.Event
+	execDone := make(chan error, 1)
+	go func() {
+		_, err := adapter.Exec(context.Background(), session, textPrompt("choose renderer"), "", "turn-ask-user", func(events []activityshared.Event) {
+			mu.Lock()
+			emittedActivity = append(emittedActivity, events...)
+			mu.Unlock()
+		}, nil)
+		execDone <- err
+	}()
+
+	waitForCondition(t, func() bool {
+		snapshot := adapter.SessionState(session)
+		return snapshot.PendingInteractive != nil &&
+			snapshot.PendingInteractive.Kind == "ask-user" &&
+			snapshot.PendingInteractive.RequestID == "permission-1"
+	})
+
+	snapshot := adapter.SessionState(session)
+	if snapshot.PendingInteractive == nil {
+		t.Fatal("pending interactive = nil, want ask-user prompt")
+	}
+	if snapshot.PendingInteractive.ToolName != "AskUserQuestion" {
+		t.Fatalf("tool name = %q, want AskUserQuestion", snapshot.PendingInteractive.ToolName)
+	}
+	questions, _ := snapshot.PendingInteractive.Input["questions"].([]any)
+	if len(questions) == 0 {
+		t.Fatalf("interactive input = %#v, want questions", snapshot.PendingInteractive.Input)
+	}
+	mu.Lock()
+	events := append([]activityshared.Event(nil), emittedActivity...)
+	mu.Unlock()
+	if requested := eventsOfType(events, activityshared.EventInteractionRequested); len(requested) != 1 ||
+		requested[0].Payload.Interaction == nil || requested[0].Payload.Interaction.Kind != "question" {
+		t.Fatalf("ask-user events = %#v, want explicit question interaction.requested", events)
+	}
+	canceled, cancel := context.WithCancel(context.Background())
+	cancel()
+	if _, err := adapter.SubmitInteractive(canceled, session, SubmitInteractiveInput{
+		RoomID:         session.RoomID,
+		AgentSessionID: session.AgentSessionID,
+		TurnID:         "turn-ask-user",
+		RequestID:      "permission-1",
+		Action:         "submit",
+	}); !errors.Is(err, context.Canceled) {
+		t.Fatalf("canceled SubmitInteractive error = %v, want context canceled", err)
+	}
+	if pending := adapter.getPendingApproval(session.AgentSessionID, "turn-ask-user", "permission-1"); pending == nil || pending.disposition() != pendingInteractiveRequestStatePending {
+		t.Fatalf("pending disposition after canceled submit = %v, want pending", runtimeInteractiveDisposition(pending))
+	}
+
+	_, err := adapter.SubmitInteractive(context.Background(), session, SubmitInteractiveInput{
+		RoomID:         session.RoomID,
+		AgentSessionID: session.AgentSessionID,
+		TurnID:         "turn-ask-user",
+		RequestID:      "permission-1",
+		Action:         "submit",
+		// Canonical GUI ask-user payload: flat display list + keyed map.
+		Payload: map[string]any{
+			"answers":             []any{"Renderer A"},
+			"answersByQuestionId": map[string]any{"render-path": "Renderer A"},
+		},
+	})
+	if err != nil {
+		t.Fatalf("SubmitInteractive: %v", err)
+	}
+	if err := <-execDone; err != nil {
+		t.Fatalf("Exec after interactive submission: %v", err)
+	}
+	outcome := transport.conn.interactiveOutcome()
+	if got := asString(outcome["outcome"]); got != "submit" {
+		t.Fatalf("interactive outcome = %#v, want submit", outcome)
+	}
+	payload, _ := outcome["payload"].(map[string]any)
+	if payload == nil || payload["answersByQuestionId"] == nil {
+		t.Fatalf("interactive payload = %#v, want answersByQuestionId", outcome)
+	}
+}
+
+// Kimi Code sends session/request_permission before the matching tool_call
+// update. The first frame identifies AskUserQuestion and its selectable
+// outcomes; the next frame carries the question body. AgentGUI reads canonical
+// Interactions rather than transcript tool rows, so the adapter must join both
+// frames before publishing interaction.requested.
+func TestStandardACPAdapterJoinsKimiAskUserQuestionInputAfterPermission(t *testing.T) {
+	t.Parallel()
+
+	transport := newStandardACPTransport("Kimi Code", "kimi-session-interactive-1")
+	transport.conn.promptKind = "ask-user-after-permission"
+	toolUpdateGate := make(chan struct{})
+	transport.conn.pauseBeforeAskUserToolUpdate = toolUpdateGate
+	var releaseToolUpdate sync.Once
+	release := func() {
+		releaseToolUpdate.Do(func() {
+			close(toolUpdateGate)
+		})
+	}
+	defer release()
+	adapter := newHermesExtensionTestAdapter(transport)
+	session := standardTestSession(hermesExtensionTestProvider)
+	session.Provider = "acp:kimi-code"
+	if _, err := adapter.Start(context.Background(), session); err != nil {
+		t.Fatalf("Start: %v", err)
+	}
+	session.ProviderSessionID = "kimi-session-interactive-1"
+
+	var mu sync.Mutex
+	var emittedActivity []activityshared.Event
+	permissionStarted := make(chan struct{}, 1)
+	requested := make(chan activityshared.Event, 1)
+	execDone := make(chan error, 1)
+	go func() {
+		_, err := adapter.Exec(context.Background(), session, textPrompt("ask how I feel"), "", "turn-kimi-ask-user", func(events []activityshared.Event) {
+			mu.Lock()
+			emittedActivity = append(emittedActivity, events...)
+			mu.Unlock()
+			for _, event := range events {
+				if event.Type == activityshared.EventCallStarted &&
+					event.Payload.CallID == "interactive-ask-1" {
+					select {
+					case permissionStarted <- struct{}{}:
+					default:
+					}
+				}
+				if event.Type == activityshared.EventInteractionRequested {
+					select {
+					case requested <- event:
+					default:
+					}
+				}
+			}
+		}, nil)
+		execDone <- err
+	}()
+
+	select {
+	case <-permissionStarted:
+	case <-time.After(2 * time.Second):
+		t.Fatal("Kimi permission request was not observed before the tool update")
+	}
+	if snapshot := adapter.SessionState(session); snapshot.PendingInteractive != nil {
+		t.Fatalf(
+			"pending interactive before tool input = %#v, want incomplete prompt hidden from SessionState",
+			snapshot.PendingInteractive,
+		)
+	}
+	mu.Lock()
+	prematureRequested := eventsOfType(emittedActivity, activityshared.EventInteractionRequested)
+	mu.Unlock()
+	if len(prematureRequested) != 0 {
+		t.Fatalf("premature interaction.requested events = %#v, want none before tool input", prematureRequested)
+	}
+	release()
+
+	var interaction activityshared.Event
+	select {
+	case interaction = <-requested:
+	case <-time.After(2 * time.Second):
+		t.Fatal("AskUserQuestion interaction was not published after the tool input arrived")
+	}
+	if interaction.Payload.Interaction == nil {
+		t.Fatalf("interaction event = %#v, want canonical interaction payload", interaction)
+	}
+	questions := payloadArray(interaction.Payload.Interaction.Input["questions"])
+	if len(questions) != 1 ||
+		asString(questions[0]["id"]) != "question-1" ||
+		asString(questions[0]["question"]) != "你今天心情怎么样？" ||
+		len(payloadArray(questions[0]["options"])) != 3 ||
+		questions[0]["allowFreeText"] != false {
+		t.Fatalf("interaction questions = %#v, want one normalized option-only Kimi question", questions)
+	}
+
+	snapshot := adapter.SessionState(session)
+	if snapshot.PendingInteractive == nil ||
+		len(payloadArray(snapshot.PendingInteractive.Input["questions"])) != 1 {
+		t.Fatalf("pending interactive = %#v, want the joined Kimi question input", snapshot.PendingInteractive)
+	}
+
+	if _, err := adapter.SubmitInteractive(context.Background(), session, SubmitInteractiveInput{
+		RoomID:         session.RoomID,
+		AgentSessionID: session.AgentSessionID,
+		TurnID:         "turn-kimi-ask-user",
+		RequestID:      "permission-1",
+		Action:         "submit",
+		Payload: map[string]any{
+			"answers":             []any{"很好"},
+			"answersByQuestionId": map[string]any{"question-1": "很好"},
+		},
+	}); err != nil {
+		t.Fatalf("SubmitInteractive: %v", err)
+	}
+	select {
+	case err := <-execDone:
+		if err != nil {
+			t.Fatalf("Exec after interactive submission: %v", err)
+		}
+	case <-time.After(2 * time.Second):
+		t.Fatal("Exec did not finish after the Kimi interactive response")
+	}
+
+	mu.Lock()
+	requestedEvents := eventsOfType(emittedActivity, activityshared.EventInteractionRequested)
+	mu.Unlock()
+	if len(requestedEvents) != 1 {
+		t.Fatalf("interaction.requested count = %d, want one complete canonical interaction", len(requestedEvents))
+	}
+	outcome := transport.conn.interactiveOutcome()
+	if got := asString(outcome["outcome"]); got != "selected" {
+		t.Fatalf("Kimi ACP outcome = %#v, want selected", outcome)
+	}
+	if got := asString(outcome["optionId"]); got != "q0_opt_0" {
+		t.Fatalf("Kimi ACP outcome = %#v, want optionId q0_opt_0", outcome)
+	}
+	if payload, _ := outcome["payload"].(map[string]any); len(payload) != 0 {
+		t.Fatalf("Kimi ACP outcome = %#v, want protocol-native selected response without payload", outcome)
+	}
+	callCompletedIndex := -1
+	providerCompletedIndex := -1
+	var callCompletedOutput map[string]any
+	for index, event := range emittedActivity {
+		if event.Type == activityshared.EventCallCompleted &&
+			event.Payload.CallID == "interactive-ask-1" {
+			callCompletedIndex = index
+			callCompletedOutput = event.Payload.Output
+		}
+		if event.Type == activityshared.EventRootProviderTurnCompleted {
+			providerCompletedIndex = index
+		}
+	}
+	if callCompletedIndex < 0 ||
+		providerCompletedIndex < 0 ||
+		callCompletedIndex >= providerCompletedIndex {
+		t.Fatalf(
+			"interactive call completed index=%d provider turn completed index=%d, want local resolution first",
+			callCompletedIndex,
+			providerCompletedIndex,
+		)
+	}
+	localPayload := payloadObject(callCompletedOutput["payload"])
+	localAnswers, _ := localPayload["answers"].([]any)
+	if len(localAnswers) != 1 || asString(localAnswers[0]) != "很好" {
+		t.Fatalf("local completed output = %#v, want the canonical answer preserved", callCompletedOutput)
+	}
+	localAnswersByQuestionID := payloadObject(localPayload["answersByQuestionId"])
+	if len(localAnswersByQuestionID) != 1 || asString(localAnswersByQuestionID["question-1"]) != "很好" {
+		t.Fatalf("local completed output = %#v, want the canonical per-question answer preserved", callCompletedOutput)
+	}
+}
+
+func TestStandardACPAdapterRejectsUnsupportedKimiAskUserQuestionBeforePublication(t *testing.T) {
+	t.Parallel()
+
+	tests := []struct {
+		name       string
+		promptKind string
+		wantError  string
+	}{
+		{
+			name:       "multiple questions",
+			promptKind: "ask-user-after-permission-multi-question",
+			wantError:  "exactly one question",
+		},
+		{
+			name:       "multi select",
+			promptKind: "ask-user-after-permission-multi-select",
+			wantError:  "does not support multi-select",
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			t.Parallel()
+
+			transport := newStandardACPTransport("Kimi Code", "kimi-session-unsupported-"+strings.ReplaceAll(tt.name, " ", "-"))
+			transport.conn.promptKind = tt.promptKind
+			adapter := newHermesExtensionTestAdapter(transport)
+			session := standardTestSession(hermesExtensionTestProvider)
+			session.Provider = "acp:kimi-code"
+			if _, err := adapter.Start(context.Background(), session); err != nil {
+				t.Fatalf("Start: %v", err)
+			}
+			session.ProviderSessionID = transport.conn.sessionID
+
+			var mu sync.Mutex
+			var emittedActivity []activityshared.Event
+			execDone := make(chan error, 1)
+			go func() {
+				_, err := adapter.Exec(context.Background(), session, textPrompt("ask unsupported question"), "", "turn-kimi-unsupported", func(events []activityshared.Event) {
+					mu.Lock()
+					emittedActivity = append(emittedActivity, events...)
+					mu.Unlock()
+				}, nil)
+				execDone <- err
+			}()
+
+			select {
+			case err := <-execDone:
+				if err != nil {
+					t.Fatalf("Exec after unsupported question rejection: %v", err)
+				}
+			case <-time.After(2 * time.Second):
+				t.Fatal("Exec did not finish after rejecting the unsupported Kimi question")
+			}
+
+			mu.Lock()
+			requested := eventsOfType(emittedActivity, activityshared.EventInteractionRequested)
+			superseded := eventsOfType(emittedActivity, activityshared.EventInteractionSuperseded)
+			completed := eventsOfType(emittedActivity, activityshared.EventCallCompleted)
+			failed := eventsOfType(emittedActivity, activityshared.EventCallFailed)
+			mu.Unlock()
+			if len(requested) != 0 {
+				t.Fatalf("interaction.requested events = %#v, want none for unsupported provider shape", requested)
+			}
+			if len(superseded) != 0 {
+				t.Fatalf("interaction.superseded events = %#v, want no canonical Interaction before publication", superseded)
+			}
+			for _, event := range completed {
+				if event.Payload.CallID == "interactive-ask-1" {
+					t.Fatalf("unsupported AskUserQuestion completed locally: %#v", event)
+				}
+			}
+			if len(failed) == 0 {
+				t.Fatal("unsupported AskUserQuestion did not emit call.failed")
+			}
+			responseErr := transport.conn.interactiveError()
+			if responseErr == nil || !strings.Contains(responseErr.Message, tt.wantError) {
+				t.Fatalf("provider response error = %#v, want containing %q", responseErr, tt.wantError)
+			}
+			if snapshot := adapter.SessionState(session); snapshot.PendingInteractive != nil {
+				t.Fatalf("pending interactive = %#v, want unsupported request removed", snapshot.PendingInteractive)
+			}
+		})
+	}
+}
+
+func TestStandardACPAdapterRejectsNonCanonicalKimiAskUserAnswer(t *testing.T) {
+	t.Parallel()
+
+	transport := newStandardACPTransport("Kimi Code", "kimi-session-invalid-answer")
+	transport.conn.promptKind = "ask-user-after-permission"
+	adapter := newHermesExtensionTestAdapter(transport)
+	session := standardTestSession(hermesExtensionTestProvider)
+	session.Provider = "acp:kimi-code"
+	if _, err := adapter.Start(context.Background(), session); err != nil {
+		t.Fatalf("Start: %v", err)
+	}
+	session.ProviderSessionID = "kimi-session-invalid-answer"
+
+	var mu sync.Mutex
+	var emittedActivity []activityshared.Event
+	requested := make(chan struct{}, 1)
+	execDone := make(chan error, 1)
+	go func() {
+		_, err := adapter.Exec(context.Background(), session, textPrompt("ask how I feel"), "", "turn-kimi-invalid-answer", func(events []activityshared.Event) {
+			mu.Lock()
+			emittedActivity = append(emittedActivity, events...)
+			mu.Unlock()
+			for _, event := range events {
+				if event.Type == activityshared.EventInteractionRequested {
+					select {
+					case requested <- struct{}{}:
+					default:
+					}
+				}
+			}
+		}, nil)
+		execDone <- err
+	}()
+
+	select {
+	case <-requested:
+	case <-time.After(2 * time.Second):
+		t.Fatal("supported Kimi question was not published")
+	}
+	result, err := adapter.SubmitInteractive(context.Background(), session, SubmitInteractiveInput{
+		RoomID:         session.RoomID,
+		AgentSessionID: session.AgentSessionID,
+		TurnID:         "turn-kimi-invalid-answer",
+		RequestID:      "permission-1",
+		Action:         "submit",
+		Payload: map[string]any{
+			"answers": []any{"很好", "一般"},
+			"answersByQuestionId": map[string]any{
+				"question-1": "很好",
+				"question-2": "一般",
+			},
+		},
+	})
+	if err == nil || !strings.Contains(err.Error(), "exactly one canonical answer") {
+		t.Fatalf("SubmitInteractive error = %v, want canonical-answer rejection", err)
+	}
+	if result.Disposition != InteractiveDispositionSuperseded {
+		t.Fatalf("SubmitInteractive disposition = %q, want superseded", result.Disposition)
+	}
+	select {
+	case err := <-execDone:
+		if err != nil {
+			t.Fatalf("Exec after invalid answer rejection: %v", err)
+		}
+	case <-time.After(2 * time.Second):
+		t.Fatal("Exec did not finish after invalid answer rejection")
+	}
+
+	mu.Lock()
+	completed := eventsOfType(emittedActivity, activityshared.EventCallCompleted)
+	failed := eventsOfType(emittedActivity, activityshared.EventCallFailed)
+	superseded := eventsOfType(emittedActivity, activityshared.EventInteractionSuperseded)
+	mu.Unlock()
+	for _, event := range completed {
+		if event.Payload.CallID == "interactive-ask-1" {
+			t.Fatalf("invalid AskUser answer completed locally: %#v", event)
+		}
+	}
+	if len(failed) == 0 || len(superseded) == 0 {
+		t.Fatalf("failed events = %d, superseded events = %d, want both", len(failed), len(superseded))
+	}
+	if responseErr := transport.conn.interactiveError(); responseErr == nil ||
+		!strings.Contains(responseErr.Message, "exactly one canonical answer") {
+		t.Fatalf("provider response error = %#v, want canonical-answer rejection", responseErr)
+	}
+	if disposition := adapter.InteractiveDisposition(session, "turn-kimi-invalid-answer", "permission-1"); disposition != InteractiveDispositionSuperseded {
+		t.Fatalf("terminal disposition = %q, want superseded", disposition)
+	}
+}
+
+func TestStandardACPAdapterSessionStateExposesPendingExitPlanPrompt(t *testing.T) {
+	t.Parallel()
+
+	transport := newStandardACPTransport("Hermes Agent", "hermes-session-plan-1")
+	transport.conn.promptKind = "exit-plan"
+	adapter := newHermesExtensionTestAdapter(transport)
+	session := standardTestSession(hermesExtensionTestProvider)
+	if _, err := adapter.Start(context.Background(), session); err != nil {
+		t.Fatalf("Start: %v", err)
+	}
+	session.ProviderSessionID = "hermes-session-plan-1"
+
+	execDone := make(chan error, 1)
+	go func() {
+		_, err := adapter.Exec(context.Background(), session, textPrompt("review plan"), "", "turn-plan", func([]activityshared.Event) {}, nil)
+		execDone <- err
+	}()
+
+	waitForCondition(t, func() bool {
+		snapshot := adapter.SessionState(session)
+		return snapshot.PendingInteractive != nil &&
+			snapshot.PendingInteractive.Kind == "exit-plan" &&
+			snapshot.PendingInteractive.RequestID == "permission-1"
+	})
+
+	snapshot := adapter.SessionState(session)
+	if snapshot.PendingInteractive == nil || snapshot.PendingInteractive.ToolName != "ExitPlanMode" {
+		t.Fatalf("pending interactive = %#v, want ExitPlanMode", snapshot.PendingInteractive)
+	}
+
+	_, err := adapter.SubmitInteractive(context.Background(), session, SubmitInteractiveInput{
+		RoomID:         session.RoomID,
+		AgentSessionID: session.AgentSessionID,
+		TurnID:         "turn-plan",
+		RequestID:      "permission-1",
+		Action:         "allow",
+		OptionID:       "acceptEdits",
+	})
+	if err != nil {
+		t.Fatalf("SubmitInteractive: %v", err)
+	}
+	if err := <-execDone; err != nil {
+		t.Fatalf("Exec after exit-plan submission: %v", err)
+	}
+	outcome := transport.conn.interactiveOutcome()
+	if got := asString(outcome["optionId"]); got != "acceptEdits" {
+		t.Fatalf("interactive outcome = %#v, want optionId acceptEdits", outcome)
+	}
+}
